Add task 4 to 14
task_4:function that frees a listint_t list.
task_5:function that frees a listint_t list
task_6:function that deletes the head node of and returns
	the head node’s data
task_7: function that returns the nth node of a listint_t
task_8:  function that returns the sum of all the data (n) of
	 a listint_t Linked list
task_9: function that inserts a new node at a given position
task_10: function that deletes the node at a given index
task_11: function that reverses a listint_t linked list.
task_12: function that prints a listint_t linked list
task_13:  function that frees a listint_t list
task_14: function that finds the loop in a linked list